pyproject.toml
[project]
name = "lunabot"
version = "0.1.0"
description = "Bot for Moonlit Lilacs Discord server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp-cors>=0.8.1",
    "asyncpg>=0.30.0",
    "dateparser>=1.2.2",
    "discord-ext-duck",
    "discord-ext-menus>=1.1",
    "discord-py>=2.6.4",
    "fuzzy-rust",
    "jinja2>=3.1.6",
    "jishaku>=2.6.3",
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "num2words>=0.5.14",
    "prettytable>=3.16.0",
    "pytz>=2025.2",
    "rapidfuzz>=3.14.1",
]

[tool.uv.sources]
discord-ext-duck = { git = "https://github.com/DuckBot-Discord/ext.duck.git" }
fuzzy-rust = { path = "fuzzy_rust" }
